initial import
updated example

HERE API 2.5.4 is used by default

Thanks to updates in the most recent HERE API it is now possible to
consume its 2.5.4 version, in which the IE10 native browser mode issue
does not appear.

The example and the docs have been updated.

Adding tiled HERE layers with the BasemapGallery dijit

The HERE tiled content may be consumed via REST API with the
WebTiledService class available in the latest versions of the ArcGIS API
for JavaScript.

readme updated

note for the BasemapGallery example: needs to be run on a website